subsections: adds 'active' and 'in-view' classes
This commit adds the 'in-view' class to a subsection if it is in the viewport,
and adds the 'active' class to the first sub-section that is in the viewport.

This is done by re-factoring the code that adds the offset data for sections
and the code that checks if a section is in the viewport to a separate function
and invoking these functions on subsections as well as sections.